Add files via upload
I didn't get the new game button working, but I implemented themes and a random one is chosen at startup. Also didn't include the label for the name of the theme. I didn't finish because I couldn't figure out how to load a new theme because I'm still unsure about the execution flow of the code when loading views and such